+++ SpeechResponder/Personality.cs
@@ -105,7 +105,8 @@ namespace EddiSpeechResponder
             "Power commodity fast tracked", // Made obsolete in Powerplay 2.0 which no longer allows fast tracking commodity allocations
             "Power defected", // Made obsolete in Powerplay 2.0 which no longer includes a defection mechanic.
             "Power expansion vote cast", // Made obsolete in Powerplay 2.0 which no longer includes a voting system
-            "Power preparation vote cast" // Made obsolete in Powerplay 2.0 which no longer includes a voting system
+            "Power preparation vote cast", // Made obsolete in Powerplay 2.0 which no longer includes a voting system
+            "Power salary claimed", // Made obsolete in Powerplay 2.0 which no longer includes a weekly salary
         };
 
         private static readonly string[] ignoredEventKeys =